// frontend/src/api/models/handlerSupplierListResponseStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerSupplierListResponseStatus =
  (typeof HandlerSupplierListResponseStatus)[keyof typeof HandlerSupplierListResponseStatus]

export const HandlerSupplierListResponseStatus = {
  active: 'active',
  inactive: 'inactive',
  blocked: 'blocked',
} as const