// frontend/src/api/models/getPartnerSuppliersParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { GetPartnerSuppliersOrderDir } from './getPartnerSuppliersOrderDir'
import type { GetPartnerSuppliersStatus } from './getPartnerSuppliersStatus'
import type { GetPartnerSuppliersType } from './getPartnerSuppliersType'

export type GetPartnerSuppliersParams = {
  /**
   * Search term (name, code, phone, email)
   */
  search?: string
  /**
   * Supplier status
   */
  status?: GetPartnerSuppliersStatus
  /**
   * Supplier type
   */
  type?: GetPartnerSuppliersType
  /**
   * City
   */
  city?: string
  /**
   * Province
   */
  province?: string
  /**
   * Minimum rating (0-5)
   */
  min_rating?: number
  /**
   * Maximum rating (0-5)
   */
  max_rating?: number
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
  /**
   * Order by field
   */
  order_by?: string
  /**
   * Order direction
   */
  order_dir?: GetPartnerSuppliersOrderDir
}