// frontend/src/api/models/getPartnerSuppliersStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetPartnerSuppliersStatus =
  (typeof GetPartnerSuppliersStatus)[keyof typeof GetPartnerSuppliersStatus]

export const GetPartnerSuppliersStatus = {
  active: 'active',
  inactive: 'inactive',
  blocked: 'blocked',
} as const