// frontend/src/api/models/getPartnerSuppliersType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetPartnerSuppliersType =
  (typeof GetPartnerSuppliersType)[keyof typeof GetPartnerSuppliersType]

export const GetPartnerSuppliersType = {
  manufacturer: 'manufacturer',
  distributor: 'distributor',
  retailer: 'retailer',
  service: 'service',
} as const